Switch to using raw HttpClientHandler as it now fully supports certificates
